Steam: warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@9399 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Steam/src/Steam/friend.h
@@ -62,7 +62,7 @@ namespace SteamWebApi
 
 		public:
 			size_t GetItemCount() const { return items.size(); }
-			const Summary *GetAt(int idx) const { return items.at(idx); }
+			const Summary *GetAt(size_t idx) const { return items.at(idx); }
 		};
 
 		static void LoadSummaries(HANDLE hConnection, const char *token, const char *steamIds, Summaries *summaries)
